% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AP}
\alias{AP}
\title{Projection matrices P and A}
\usage{
AP(x, v, ind_dj, orth = FALSE)
}
\arguments{
\item{x, v}{row vectors of size \code{sum(d) + r}.}

\item{ind_dj}{\code{0}-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}

\item{orth}{return the orthogonal complement of \eqn{P}, \eqn{I - P}?}
}
\description{
The \eqn{jj}-block of \eqn{P} is \eqn{I_{d_j} - x_j x_j'}. The
\eqn{jj}-block of \eqn{A} is \eqn{(x_j' * v_j) * I_{d_j}}.
}
\keyword{internal}
