% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{block_euler_ridge}
\alias{block_euler_ridge}
\title{Blockwise Euler algorithm for density ridge estimation}
\usage{
block_euler_ridge(x, X, d, h, h_euler, ind_blocks, N = 1000, eps = 1e-05,
  keep_paths = FALSE, cores = 1, ...)
}
\arguments{
\item{x}{a matrix of size \code{c(nx, sum(d) + r)} with the evaluation
points.}

\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{d}{dimensions vector, of size \code{r}.}

\item{h}{bandwidth vector, of size \code{r}.}

\item{h_euler}{vector of size \code{r} with the advance steps in the Euler
method. Set internally as \code{h} if not provided.}

\item{ind_blocks}{indexes of the blocks, a vector or length \code{r}.}

\item{N}{maximum number of Euler iterations. Defaults to \code{1e3}.}

\item{eps}{convergence tolerance. Defaults to \code{1e-5}.}

\item{keep_paths}{keep the Euler paths to the ridge? Defaults to
\code{FALSE}.}

\item{cores}{cores to use. Defaults to \code{1}.}

\item{...}{further parameters passed to \code{\link{euler_ridge}}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
