% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{c_kern}
\alias{c_kern}
\title{Normalizing constant of the a directional kernel}
\usage{
c_kern(h, d, kernel = "1", kernel_type = "1", k = 10, log = FALSE,
  inc_sfp = TRUE, intrinsic = FALSE)
}
\arguments{
\item{h}{bandwidth vector, of size \code{r}.}

\item{d}{dimensions vector, of size \code{r}.}

\item{kernel}{kernel employed: \code{1} for vMF (default); \code{2}
for Epa, \eqn{L(t) = (1 - t)1_{\{0 \le t \le 1\}}}; \code{3} for softplus.}

\item{kernel_type}{type of kernel employed: \code{1} for product kernel
(default); \code{2} for spherically symmetric kernel.}

\item{k}{softplus kernel parameter. Defaults to \code{10.0}.}

\item{log}{compute the logarithm of the constant?}

\item{inc_sfp}{include the softplus(k) function in the constant? Defaults to
\code{TRUE}.
\code{TRUE}.}

\item{intrinsic}{consider the intrinsic distance? Defaults to \code{FALSE}.}
}
\description{
Normalizing constant of the a directional kernel
}
\details{
This function computes the normalizing constant \eqn{c(h,d,L)}
such that \eqn{\int_{S^d} c(h,d,L) L((1 - x'y)/h^2) dx = 1}
(extrinsic-chordal distance) or
\eqn{\int_{S^d} c(h,d,L) L(arccos(x'y)^2/(2h^2)) dx = 1} (intrinsic
distance).
}
\examples{
h_grid <- seq(0.01, 4, l = 100)
r <- 2
d <- 2
dr <- rep(d, r)
c_vmf <- sapply(h_grid, function(hi)
  log(c_kern(h = rep(hi, r), d = dr, kernel = 1, kernel_type = 2)))
c_epa <- sapply(h_grid, function(hi)
  log(c_kern(h = rep(hi, r), d = dr, kernel = 2, kernel_type = 2)))
c_sfp <- sapply(h_grid, function(hi)
  log(c_kern(h = rep(hi, r), d = dr, kernel = 3, k = 1, kernel_type = 2)))
plot(h_grid, c_epa, type = "l", ylab = "Constant", xlab = "h", col = 2)
lines(h_grid, c_sfp, col = 3)
lines(h_grid, c_vmf, col = 1)
abline(v = sqrt(2), lty = 2, col = 2)
}
