% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwd.R
\name{curv_vmf_polysph}
\alias{curv_vmf_polysph}
\title{Curvature of a polyspherical von Mises--Fisher density}
\usage{
curv_vmf_polysph(kappa, d, log = FALSE)
}
\arguments{
\item{kappa}{a vector of size \code{r} with the vMF concentrations.}

\item{d}{dimensions vector, of size \code{r}.}

\item{log}{compute the logarithm of the curvature? Defaults to \code{FALSE}.}
}
\description{
Computes the curvature matrix \eqn{\boldsymbol{R}} of a
polyspherical von Mises--Fisher density.
}
