% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{d_unif_polysph}
\alias{d_unif_polysph}
\title{TODO}
\usage{
d_unif_polysph(x, d, log = FALSE)
}
\arguments{
\item{x}{TODO}

\item{d}{TODO}

\item{log}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
