% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{d_vmf_polysph}
\alias{d_vmf_polysph}
\title{TODO}
\usage{
d_vmf_polysph(x, d, mu, kappa, log = FALSE)
}
\arguments{
\item{x}{TODO}

\item{d}{TODO}

\item{mu}{TODO}

\item{kappa}{TODO}

\item{log}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
