% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diamond_crossprod}
\alias{diamond_crossprod}
\title{Compute cube \eqn{X_i \diamond X_i'}}
\usage{
diamond_crossprod(X, ind_dj)
}
\arguments{
\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{ind_dj}{\code{0}-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
\keyword{internal}
