% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_polysph_matrix}
\alias{dist_polysph_matrix}
\title{Polyspherical distance matrix}
\usage{
dist_polysph_matrix(x, ind_dj, norm_x = FALSE, norm_y = FALSE,
  std = TRUE)
}
\arguments{
\item{x}{a matrix of size \code{c(n, sum(d) + r)}.}

\item{ind_dj}{0-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}

\item{norm_x, norm_y}{ensure a normalization of the data?}

\item{std}{standardize distance to \eqn{[0,1]}? Uses that the maximum
distance is \eqn{\sqrt{r}\pi}. Defaults to \code{TRUE}.}
}
\description{
Polyspherical distance matrix
}
\examples{
# Example on S^2 x S^3 x S^1
d <- c(2, 3, 1)
ind_dj <- comp_ind_dj(d)
x <- r_unif_polysph(n = 5, d = d)
dd <- dist_polysph_matrix(x, ind_dj = ind_dj, std = FALSE)
dd

# Simple check
i <- 1
j <- 3
dd[i, j]
sqrt(acos(x[i, 1:3] \%*\% x[j, 1:3])^2 + acos(x[i, 4:7] \%*\% x[j, 4:7])^2 +
       acos(x[i, 8:9] \%*\% x[j, 8:9])^2)
}
