% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{hd_mc}
\alias{hd_mc}
\title{Hellinger distance via Monte Carlo}
\usage{
hd_mc(log_f, log_g, bhatta = FALSE)
}
\arguments{
\item{log_f, log_g}{logarithms of the pdfs f and g evaluated in a Monte Carlo
sample for approximating the integral \eqn{\int \sqrt{f(x) * g(x)} dx}.}

\item{bhatta}{compute the Bhattacharyya divergence (or coefficient) instead?}
}
\description{
Hellinger distance via Monte Carlo
}
