% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sreps.R
\name{interp_srep}
\alias{interp_srep}
\title{Movie weighted by distances}
\usage{
interp_srep(X, ind_dj, ord, L = NULL)
}
\arguments{
\item{X}{TODO}

\item{ind_dj}{0-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}

\item{ord}{TODO}

\item{L}{TODO}
}
\description{
Movie weighted by distances
}
