% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{J_d_k}
\alias{J_d_k}
\title{Computes the integral \eqn{J_{d, k}}}
\usage{
J_d_k(d, k = 10, upper = Inf, ...)
}
\arguments{
\item{d}{either a vector with the dimensions or a scalar with the
common dimension of each hypersphere (a scalar).}

\item{k}{softplus kernel parameter. Defaults to \code{10.0}.}

\item{upper}{upper limit of integration.}

\item{...}{further parameters passed to \code{\link{integrate}}, such as
\code{upper}, \code{abs.tol}, \code{rel.tol}, etc.}
}
\value{
TODO
}
\description{
Computes the integral \eqn{J_{d, k}=TODO}.
}
\examples{
# TODO
}
\keyword{internal}
