% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{L}
\alias{L}
\alias{grad_L}
\alias{hess_L}
\title{Directional kernels and their derivatives}
\usage{
L(t, kernel = "1", squared = FALSE, deriv = 0, k = 10,
  inc_sfp = TRUE)

grad_L(x, y, h, kernel = 1, k = 10)

hess_L(x, y, h, kernel = 1, k = 10)
}
\arguments{
\item{t}{kernel argument. A positive scalar.}

\item{kernel}{kernel employed: \code{1} for vMF (default); \code{2}
for Epa, \eqn{L(t) = (1 - t)1_{\{0 \le t \le 1\}}}; \code{3} for softplus.}

\item{squared}{square the kernel? Only for \code{deriv == 0}. Defaults to
\code{FALSE}.}

\item{deriv}{kernel derivative. Must be \code{0}, \code{1}, or \code{2}.
Defaults to \code{0}.}

\item{k}{softplus kernel parameter. Defaults to \code{10.0}.}

\item{inc_sfp}{include \code{softplus(k)} in the constant? Defaults to
\code{TRUE}.}

\item{x}{a matrix of size \code{c(nx, sum(d) + r)} with the evaluation
points.}

\item{y}{center of the kernel.}

\item{h}{bandwidth vector, of size \code{r}.}
}
\description{
Directional kernels and their derivatives
}
\details{
TODO
}
