% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Safe evaluation of \eqn{\log(\sum_{i=1}^n e^{x_i})}}
\usage{
log_sum_exp(logs, avg = FALSE)
}
\arguments{
\item{logs}{vector with \eqn{x_i} values.}

\item{avg}{replace the sum by the average? Defaults to \code{FALSE}.}
}
\value{
A vector of size \code{length(x)} with the evaluated function.
}
\description{
Computes \eqn{\log(\sum_{i=1}^n e^{x_i})} using the "LogSumExp
trick".
}
\examples{
logs <- c(1e5, 1)
log(sum(exp(logs)))
polykde:::log_sum_exp(logs)
}
