% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{polysph_to_angles}
\alias{polysph_to_angles}
\alias{angles_to_polysph}
\title{Conversion between the angular and Cartesian coordinates of the
polysphere}
\usage{
polysph_to_angles(x, dims)

angles_to_polysph(theta, dims)
}
\arguments{
\item{x}{matrix of size \code{c(n, sum(dims + 1))} with the Cartesian
coordinates.}

\item{dims}{vector with the dimensions of the polysphere.}

\item{theta}{matrix of size \code{c(n, sum(dims))} with the angles.}
}
\description{
Conversion between the angular and Cartesian coordinates of the
polysphere
}
\examples{
polysph_to_angles(x = c(0, 0, 1, 0, 1), dims = c(2, 1))
angles_to_polysph(rep(pi / 2, 3), dims = c(2, 1))
}
