% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s}
\alias{s}
\title{Symmetrize a matrix}
\usage{
s(A, add = FALSE)
}
\arguments{
\item{A}{a matrix.}

\item{add}{return simply the addition
\eqn{\boldsymbol{A} + \boldsymbol{A}'}? Defaults to \code{FALSE}}
}
\value{
A symmetric matrix with the same dimensions as \code{A}.
}
\description{
Symmetrizes a matrix \eqn{\boldsymbol{A}} by returning
\eqn{(\boldsymbol{A} + \boldsymbol{A}') / 2}.
}
\examples{
A <- matrix(rnorm(4), nrow = 2, ncol = 2)
polykde:::s(A)
}
\keyword{internal}
