% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfp}
\alias{sfp}
\title{Stable computation of the softplus function}
\usage{
sfp(t)
}
\arguments{
\item{t}{numeric vector or matrix.}
}
\value{
The softplus function evaluated at \code{t}.
}
\description{
Computes the softplus function \eqn{\log(1+e^{t})} in a
numerically stable way for large absolute values of \eqn{t}.
}
\examples{
curve(log(polykde:::sfp(rbind(5 * (1 - x)))), from = -10, to = 10)
}
\keyword{internal}
