% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sreps.R
\name{view_srep_movie}
\alias{view_srep_movie}
\title{View s-reps across time}
\usage{
view_srep_movie(avg_base, avg_radii, avg_dirs, moving_spokes, X, ord,
  col_pal = grDevices::rainbow, L = NULL, static = TRUE, ...)
}
\arguments{
\item{avg_base}{TODO}

\item{avg_radii}{TODO}

\item{avg_dirs}{TODO}

\item{moving_spokes}{TODO}

\item{X}{TODO}

\item{ord}{TODO}

\item{col_pal}{TODO}

\item{L}{TODO}

\item{static}{TODO}

\item{...}{TODO}
}
\description{
View s-reps across time
}
\keyword{internal}
