% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AP}
\alias{AP}
\title{Projection matrices \eqn{\boldsymbol{P}} and \eqn{\boldsymbol{A}}}
\usage{
AP(x, v, ind_dj, orth = FALSE)
}
\arguments{
\item{x, v}{row vectors of size \code{sum(d) + r}.}

\item{ind_dj}{\code{0}-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}

\item{orth}{return the orthogonal complement of \eqn{P}, \eqn{I - P}?}
}
\value{
TODO
}
\description{
TODO

The \eqn{jj}-block of \eqn{\boldsymbol{P}} is
\eqn{\boldsymbol{I}_{d_j} - \boldsymbol{x}_j \boldsymbol{x}_j'}. The
\eqn{jj}-block of \eqn{\boldsymbol{A}} is
\eqn{(\boldsymbol{x}_j' \boldsymbol{v}_j) \boldsymbol{I}_{d_j}}.
}
\examples{
# TODO
}
\keyword{internal}
