% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{beta0_R}
\alias{beta0_R}
\title{Fixed-point algorithm for the weighted Epanechnikov kernel}
\usage{
beta0_R(d, R, tol = 1e-10)
}
\arguments{
\item{d}{either a vector with the dimensions or a scalar with the
common dimension of each hypersphere (a scalar).}

\item{R}{curvature matrix as outputted by \code{\link{curv_vmf_polysph}}.}

\item{tol}{tolerance for the fixed-point algorithm. Defaults to \code{1e-10}.}
}
\value{
Vector \eqn{\boldsymbol{\beta}} of weights.
}
\description{
Solves the equation
\deqn{\boldsymbol{\beta} = \mathrm{diag}(\boldsymbol{d}^{\odot (-1)})
\boldsymbol{R}^{-1} \boldsymbol{\beta}^{\odot(-1)}}
to determine the weights of the weighted Epanechnikov kernel.
}
\examples{
r <- 2
d <- seq_len(r)
R <- curv_vmf_polysph(kappa = 10 * d, d = d)
polykde:::beta0_R(d = d, R = R)
}
\keyword{internal}
