% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_lists}
\alias{bind_lists}
\title{Merge a list of lists}
\usage{
bind_lists(lists, bind = "rbind")
}
\arguments{
\item{lists}{lists whose entries are to be binded.}

\item{bind}{bind operator, either \code{"rbind"} or \code{"cbind"}.}
}
\description{
Merge a list of lists
}
