% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwd.R
\name{bw_lcv_min_epa}
\alias{bw_lcv_min_epa}
\title{Minimum Epanechnikov bandwidth allowed in LCV}
\usage{
bw_lcv_min_epa(X, d, kernel_type = c("prod", "sph")[1])
}
\arguments{
\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{d}{dimensions vector, of size \code{r}.}

\item{kernel_type}{type of kernel employed: \code{1} for product kernel
(default); \code{2} for spherically symmetric kernel.}
}
\description{
Minimum Epanechnikov bandwidth allowed in LCV
}
