% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwd.R
\name{bw_rot_polysph}
\alias{bw_rot_polysph}
\title{Rule-of-thumb bandwidths}
\usage{
bw_rot_polysph(X, d, kernel = 1, kernel_type = c("prod", "sph")[1],
  bw0 = NULL, upscale = FALSE, deriv = 0, k = 10, kappa = NULL, ...)
}
\arguments{
\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{d}{dimensions vector, of size \code{r}.}

\item{kernel}{kernel employed: \code{1} for vMF (default); \code{2}
for Epa, \eqn{L(t) = (1 - t)1_{\{0 \le t \le 1\}}}; \code{3} for softplus.}

\item{kernel_type}{type of kernel employed: \code{1} for product kernel
(default); \code{2} for spherically symmetric kernel.}

\item{bw0}{initial bandwidth for minimizing the AMISE. Computed with
\code{\link{bw_rot_polysph}} if not provided (default).}

\item{upscale}{rescale bandwidths to work on
\eqn{S^{d_1}\times\cdots\times S^{d_r}} and for derivative estimation?
Defaults to \code{FALSE}. If \code{upscale == 1}, the order \code{n} is
upscaled. If \code{upscale == 2}, then also the kernel constant is upscaled.}

\item{deriv}{derivative order to perform the upscaling. Defaults to \code{0}.}

\item{k}{softplus kernel parameter. Defaults to \code{10.0}.}

\item{kappa}{estimate of the concentration parameters. Computed if not
provided (default).}

\item{...}{further parameters passed to \code{\link[=stats]{optim}}.}
}
\description{
Rule-of-thumb bandwidths
}
