% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{clean_euler_ridge}
\alias{clean_euler_ridge}
\title{Clean ridge points coming from spurious fits}
\usage{
clean_euler_ridge(e, X, p_out = NULL)
}
\arguments{
\item{e}{outcome from \code{\link{euler_ridge}} or
\code{\link{parallel_euler_ridge}}.}

\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{p_out}{proportion of outliers to remove. Defaults to \code{NULL} (no
cleaning).}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
