% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_polysph}
\alias{dist_polysph}
\title{Polyspherical distance}
\usage{
dist_polysph(x, y, ind_dj, norm_x = FALSE, norm_y = FALSE, std = TRUE)
}
\arguments{
\item{x}{a matrix of size \code{c(n, sum(d) + r)}.}

\item{y}{either a matrix of the same dimension of \code{x} or a vector of
length \code{sum(d) + r}.}

\item{ind_dj}{\code{0}-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}

\item{norm_x, norm_y}{ensure a normalization of the data?}

\item{std}{standardize distance to \eqn{[0,1]}? Uses that the maximum
distance is \eqn{\sqrt{r}\pi}. Defaults to \code{TRUE}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
