% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{index_ridge}
\alias{index_ridge}
\title{Index a ridge curve, creating the smoothed and indexed ridge}
\usage{
index_ridge(endpoints, X, d, l_index = 1000, f_index = 2,
  probs_scores = seq(0, 1, l = 101), verbose = FALSE, type_bwd = c("1se",
  "min")[1], p = 0, ...)
}
\arguments{
\item{endpoints}{end points of the ridge algorithm to be indexed.}

\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{d}{dimensions vector, of size \code{r}.}

\item{l_index}{length of the grid index used for finding projections.
Defaults to \code{1e3}.}

\item{f_index}{factor with the range of the grid for finding ridge
projections. Defaults to \code{2}, which is twice the range of MDS indexes.}

\item{probs_scores}{probabilities for indexing the ridge on the
\code{probs_scores}-quantiles of the scores. Defaults to
\code{seq(0, 1, l = 101)}.}

\item{verbose}{show diagnostic plots?}

\item{type_bwd}{type of cross-validation bandwidth for Nadaraya--Watson,
either \code{"min"} (minimizer of the cross-validation loss) or \code{"1se"}
(the "one standard error rule", smoother). Defaults to \code{"1se"}.}

\item{p}{degree of local fit, either \code{0} or \code{1}. Defaults to
\code{0}.}

\item{...}{further parameters passed to \code{\link{bw_cv_kre_polysph}}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
