% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sreps.R
\name{interp_polysph}
\alias{interp_polysph}
\title{Interpolation on the polysphere}
\usage{
interp_polysph(x, y, ind_dj, N = 1)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{ind_dj}{0-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}

\item{N}{TODO}
}
\description{
Interpolation on the polysphere
}
\keyword{internal}
