% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kre.R
\name{kre_polysph}
\alias{kre_polysph}
\alias{bw_cv_kre_polysph}
\title{Local polynomial estimator for polyspherical-on-scalar regression}
\usage{
kre_polysph(x, X, Y, d, h, p = 0)

bw_cv_kre_polysph(X, Y, d, p = 0, h_grid = bw.nrd(X) * 10^seq(-2, 2, l =
  100), plot_cv = TRUE, fast = TRUE)
}
\arguments{
\item{x}{a vector of size \code{n} with the evaluation points.}

\item{X}{a vector of size \code{n} with the predictor sample.}

\item{Y}{a matrix of size \code{c(n, sum(d) + r)} with the response sample.}

\item{d}{dimensions vector, of size \code{r}.}

\item{h}{a positive scalar giving the bandwidth.}

\item{p}{degree of local fit, either \code{0} or \code{1}. Defaults to
\code{0}.}

\item{h_grid}{bandwidth grid where to optimize the cross-validation loss.
Defaults to \code{bw.nrd(X) * 10^seq(-1, 1, l = 100)}.}

\item{plot_cv}{plot the cross-validation loss curve? Defaults to \code{TRUE}.}

\item{fast}{use the faster and equivalent version of the cross-validation
loss? Defaults to \code{TRUE}.}
}
\description{
Local polynomial estimator for polyspherical-on-scalar regression
}
