% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proj_polysph}
\alias{proj_polysph}
\title{Polyspherical projection}
\usage{
proj_polysph(x, ind_dj)
}
\arguments{
\item{x}{a matrix of size \code{c(nx, sum(d) + r)} with the evaluation
points.}

\item{ind_dj}{\code{0}-based index separating the spheres. Computed using
\code{\link{comp_ind_dj}}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
\keyword{internal}
