% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{r_g_kern}
\alias{r_g_kern}
\title{Sample from the angular kernel function}
\usage{
r_g_kern(n, d, h, kernel = "1", k = 10)
}
\arguments{
\item{n}{sample size.}

\item{d}{dimensions vector, of size \code{r}.}

\item{h}{bandwidth vector, of size \code{r}.}

\item{kernel}{kernel employed: \code{1} for vMF (default); \code{2}
for Epa, \eqn{L(t) = (1 - t)1_{\{0 \le t \le 1\}}}; \code{3} for softplus.}

\item{k}{softplus kernel parameter. Defaults to \code{10.0}.}
}
\description{
Sample from the angular kernel function
}
