% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{r_kde_polysph}
\alias{r_kde_polysph}
\title{Sample from polyspherical kde}
\usage{
r_kde_polysph(n, X, d, h, kernel = 1, kernel_type = 1, k = 10,
  norm_X = FALSE)
}
\arguments{
\item{n}{sample size.}

\item{X}{a matrix of size \code{c(n, sum(d) + r)} with the sample.}

\item{d}{dimensions vector, of size \code{r}.}

\item{h}{bandwidth vector, of size \code{r}.}

\item{kernel}{kernel employed: \code{1} for vMF (default); \code{2}
for Epa, \eqn{L(t) = (1 - t)1_{\{0 \le t \le 1\}}}; \code{3} for softplus.}

\item{kernel_type}{type of kernel employed: \code{1} for product kernel
(default); \code{2} for spherically symmetric kernel.}

\item{k}{softplus kernel parameter. Defaults to \code{10.0}.}

\item{norm_X}{ensure a normalization of the data?}
}
\description{
Sample from polyspherical kde
}
