% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{r_unif_polysph}
\alias{r_unif_polysph}
\title{Sample uniform polyspherical data}
\usage{
r_unif_polysph(n, d)
}
\arguments{
\item{n}{sample size.}

\item{d}{dimensions vector, of size \code{r}.}
}
\description{
Sample uniform polyspherical data
}
