% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{r_vmf_polysph}
\alias{r_vmf_polysph}
\title{Sample vMF-distributed polyspherical data}
\usage{
r_vmf_polysph(n, d, mu, kappa, normalize_mu = FALSE)
}
\arguments{
\item{n}{sample size.}

\item{d}{dimensions vector, of size \code{r}.}

\item{mu}{a vector of size \code{sum(d) + r} with the concatenated vMF means
on \eqn{S^{d_j}}.}

\item{kappa}{a vector of size \code{r} with the vMF concentrations.}

\item{normalize_mu}{ensure a normalization of \code{mu}?}
}
\description{
Sample vMF-distributed polyspherical data
}
