% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s}
\alias{s}
\title{Symmetrize a matrix A with (A+A')/2}
\usage{
s(A, add = FALSE)
}
\arguments{
\item{A}{matrix.}

\item{add}{return simply the addition A + A'? Defaults to \code{FALSE}}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
\keyword{internal}
