% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfp}
\alias{sfp}
\title{Stable computation of the softplus function}
\usage{
sfp(t)
}
\arguments{
\item{t}{numeric vector or matrix.}
}
\value{
TODO
}
\description{
Computes the softplus function in a numerically stable way.
}
\examples{
curve(log(polykde:::sfp(rbind(5 * (1 - x)))), from = -10, to = 10)
}
\keyword{internal}
