% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softplus}
\alias{softplus}
\title{Stable computation of the softplus function}
\usage{
softplus(t)
}
\arguments{
\item{t}{numeric vector or matrix.}
}
\description{
Stable computation of the softplus function
}
\examples{
h <- 0.1
k <- 1
curve(polykde:::softplus(k * (1 - (1 - x) / h^2)), from = -1, to = 1)
}
\keyword{internal}
