% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sreps.R
\name{view_srep}
\alias{view_srep}
\title{s-rep viewer}
\usage{
view_srep(base, dirs, bdry, radii, show_base = TRUE, show_base_pt = TRUE,
  show_bdry = TRUE, show_bdry_pt = TRUE, show_seg = TRUE,
  col_base = "red", col_bndy = "blue", col_seg = "green",
  static = FALSE, texts = NULL, cex_base = ifelse(static, 0.5, 6),
  cex_bdry = ifelse(static, 1, 8), lwd_seg = ifelse(static, 1, 2),
  cex_texts = 1, alpha_base = 0.1, alpha_bdry = 0.15, r_texts = 1.25,
  alpha_ashape3d_base = 100, alpha_ashape3d_bdry = 5, lit = FALSE, ...)
}
\arguments{
\item{base}{TODO}

\item{dirs}{TODO}

\item{bdry}{TODO}

\item{radii}{TODO}

\item{show_base, show_base_pt}{TODO}

\item{show_bdry, show_bdry_pt}{TODO}

\item{show_seg}{TODO}

\item{col_base, col_bndy, col_seg}{TODO}

\item{static}{TODO}

\item{texts}{TODO}

\item{cex_base, cex_bdry}{TODO}

\item{lwd_seg}{TODO}

\item{cex_texts}{TODO}

\item{alpha_base, alpha_bdry}{TODO}

\item{r_texts}{TODO}

\item{alpha_ashape3d_base, alpha_ashape3d_bdry}{TODO}

\item{lit}{TODO}

\item{...}{TODO}
}
\description{
s-rep viewer
}
\keyword{internal}
